AvalancheModularityLayer compiles with caveats.
The stuff in YourCodeHere currently does not compile.
Execution of this code has not been attempted.
This should probably have a lot more automated tests than it does,
especially for some of the algorithmic stuff like the InputSynchronizer
and the InputEventQueue.
Design changes are still needed to make input handling better.